requirements.txt
asgiref==3.6.0
coverage==7.2.7
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
